Fixed test launch files so that tests can be executed from the IDE again
Target files had to be cleaned up as well in the process and now
contain separate locations as well as a method how to setup Skalli
initially incorporating the suggestion from the contributer guide. One
simply needs to download the 3rd party libraries now and does not need
to adapt anything else.

Bug: 348852
Change-Id: Iba9ceb01eed816eace319fe5680df22f3b3409d3
Signed-off-by: Robert Wetzold <robert.wetzold@sap.com>
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>