StorageService
Up to now content store for project data and customization was
the local file system ("storage dir"). This patch introduces a new
service interface, StorageService, to allow alternative content
stores, such as a database. Note, we still translate our model classes
with XStream to XML, but instead of writing the result to a file
directly, we pass it to the storage service.

* introduced StorageService, which resembles a "blob store"
* added default implementation FileStorageService that uses the local
file system as content store
* adapted PersistenceServiceXStream to use the storage service (bound
via OSGi DS)
* refactoring of PersistenceServiceXStream, XStreamPersistence and
some helper classes

Bug: 349549
Change-Id: I2c000859d9d705ab0377a3201f1b8020f02318a0
Signed-off-by: Britta Varwig <britta.varwig@sap.com>
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>