Introduced StorageService for configurations
Up to now configuration entries were stored in the Eclipse secure
store which is not available for all platforms and requires a local
file system. Adapted the implementation to use the new StorageService
interface. All configuration properties are written to one single
customization entry with key "PROPERTYSTORE" now.

Drawback of this solution: We loose the encryption of the Eclipse
secure store.

Bug: 349549
Change-Id: Ifdc25717d09fc39755605f1d37ef61377ec32636
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>