Make the StorageService configurable in PersistenceServiceXStream
By setting the property "skalli.storageService" either in a
skalli.properties resource file (just add a fragment to o.e.s.core
that contains this file in the bundle's root folder!) or as system
property, it is now possible to select a dedicated StorageService.
If nothing is set, FileStorageService is choosen as fallback.

Value of the property is in both cases the qualified class name of the
storage service implementation.

Example:
    skalli.storageService=org.eclipse.skalli.storage.db.PersistenceDB
selects the database storage service for persistency.

Note: You can of course still uses the OSGi console to enable/disable
the appropriate storage service. However, the property mechanism can
be helpful on runtimes, where access to the OSGi console is limited.

Bug: 350793
Change-Id: I1b6884356eb9ec5d1e3b1dbfd1689d20c98214f8
Signed-off-by: Britta Varwig <britta.varwig@sap.com>
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>