New CustomField implementation breaks MultiLinkField
Delegation does not work if field derived from CustomField
overrides getValue(): commit internally calls getValue(),
which was delegated to AbstractField, but it should call
the overridden method instead.
That is a fundamental problem with the delegator
solution, but the only alternative would mean to copy
large amount of code from AbstractField --- which we are
not allowed to.

Change-Id: Iba3ae5be51844f0d380756286696bcaad1cc9bd6